+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix Performance issue on v2 Flaw endpoint with include/exclude queryparams.
 - Added lost indexes for AffectsV1 and tests to ensure they are not deleted (OSIDB-4717).
+- Allow updates for flaw labels in any workflow state (OSIDB-4737)
 
 ## Removed
 - Remove CWE ID collection from collectors (OSIDB-4742)